http client fix
Summary:
There was an upstream change to the rust-client lib that required changes to compile.
The lib now requires the body be `Serialize` and not just `Deref<IngestBody>`.
This conforms to these changes.

Test Plan:
Run a couple million lines through locally and been on stage for almost a day.
Body serialization seems to be working as expected.

Reviewers: #rust, jakedipity

Reviewed By: #rust, jakedipity

Differential Revision: https://logdna.phacility.com/D1807