switch to single-threaded critical path
Summary:
This removes all threading from the critical path and now uses a single tight loop.
Functionality wise things are relatively unchanged, there is a slight difference in how we poll inotify so we don't block for too long.

The main motivation for this was to reduce the overhead of the synchronization(message passing) in the critical path and reduce the memory footprint by stripping away the threads.

The agent's working set is now around 300KB, which 200KB of which k8s meta (labels/annotations).

Test Plan:
Baked a few days on stage.
Saw no notable difference in log volume over the few days.
Ran 10 rounds of 1 mill lines.
Used valgrind(callgrind, cachegrind, massif) and heaptrack to measure the impact of the changes.

Reviewers: #rust, dchai

Reviewed By: #rust, dchai

Subscribers: dchai

Differential Revision: https://logdna.phacility.com/D1878