Add journald hook
Summary:
This diff aims to hook into journald and use it as a source for log lines. This change also standardizes a way for multiple sources to feed into the main event cycle of the logdna-agent through a new `Source` trait and `SourceReader` class. This diff is still rough around the edges, but I want to get some input before getting to deep into other parts of this implementation.

https://app.asana.com/0/0/1156270392649081/f

Test Plan:
Tested that journald generated events create a log event: `echo "test" | systemd-cat`.

Also tested that the original file system tail works as expected.

Built and tested on arch linux.

Reviewers: #integrations, #rust, connor.peticca

Reviewed By: #rust, connor.peticca

Subscribers: connor.peticca

Differential Revision: https://logdna.phacility.com/D2782