Add k8s event stream log source
LOG-7719

Add a new log source that consumes kubernetes events from the Events api
and forwards to logdna.

In order to avoid excessive event duplication the oldest agent of the
current generation is responsible for uploading the events. All other
agents watch for the eldest agent to be deleted at which point the next
agent that meets the criteria will start logging events.
In the case that one or more agents in a generation is the same age the
first alphabetically will upload the events.