Fix bug where Journald ingests stale records
Found an issue with Journald that occurs in the following scenario

1. The agent is already started and monitoring a journal such as `/var/log/journal`
2. The node does not currently have Journald logging enabled
3. Modifying the node to enable journald which results in the creation of a new subdirectory `/var/log/journal/<GUID>`
4. The agent will start going through all the stale journald records from the beginning and giving them the false timestamp of `now`

This issue occurs because we only seek the tail of the Journal stream on startup. This fix checks to make sure all records have a timestamp of within 30 seconds. If a timestamp is older than that, then the journal is reseeked to the end.

```
[2020-11-10T10:10:52Z WARN  journald::stream] Received a stale journald record, reseeking pointer
```

Signed-off-by: Jacob Hull <jacob@planethull.com>