K8s event stream (#37)
* Update k8s metadata to use kube-runtime watcher

* Add k8s event stream log source

LOG-7719

Add a new log source that consumes kubernetes events from the Events api
and forwards to logdna.

In order to avoid excessive event duplication the oldest agent of the
current generation is responsible for uploading the events. All other
agents watch for the eldest agent to be deleted at which point the next
agent that meets the criteria will start logging events.
In the case that one or more agents in a generation is the same age the
first alphabetically will upload the events.

* Make things look better for the UI

1. The `app` was incorrectly being parsed from the `field_path`
2. Fixed a bug where Events with a single occurrence didn't get a duration because the start and last timestamp were exactly the same
3. Fixed a case where some events don't have a duration, age, or count because they can only ever represent 1 occurrence

e.g.
```
NAMESPACE      LAST SEEN   TYPE      REASON              OBJECT                   MESSAGE
logdna-agent   <unknown>   Normal    Scheduled           pod/logdna-agent-bmdj5   Successfully assigned logdna-agent/logdna-agent-bmdj5 to minikube
```

For these cases we don't set `first_time` and set `time` to the time of the event. We also modify the message to not include the `(x2 13 seconds)`

Doesn't seem like Kube Reporter handles these cases

4. Since the we can directly control the line we send the agents we don't need to use the weird host/app/level workaround; just set it directly in the request

5. Fix an issue where the `kube.type` was incorrectly sent as `kube.type_`

Signed-off-by: Jacob Hull <jacob@planethull.com>

Co-authored-by: Jacob Hull <jacob@planethull.com>