fix: delay k8s metadata cleanup until file is deleted
In some cases, logs ingested closed to when a pod is deleted might cause
the logs to not be able to collect metadata information since the info
has been removed from our cache. This fix makes sure that metadata info
lives as long as the file exists to prevent this from happening.

Signed-off-by: Jacob Hull <jacob@planethull.com>