fix: silently replace existing fs entries
Not silently replacing fs entries, and instead returning an error, short
circuits the logic that normally handles processing new files.

In cases where log rotation re-uses the same inodes, log rotation is
seen as log truncation. Truncation is only detected during write
events. If this event surpasses the internal 8196 byte limit, tailing
will be picked up at the end of the file. This causes the first set of
lines to be dropped.

By fixing this short circuiting issue, log truncation should now be
correctly handled once the file is created. This should appropriately
reset the offsets of the tailed file so the next write event begins
properly at the beginning.

Signed-off-by: Jacob Hull <jacob@planethull.com>