fix: handle truncation while assigning initial offsets
Initial offsets never considered the current state of the file, instead
just preferring to use the state saved offsets or defaulting to the
value configured by the `MZ_LOOKBACK` option. This means truncation is
handled downstream where the agent doesn't have enough contextual
information to appropriately handle the case and potentially dropping
logs.

This change causes initial offsets to be ignored if they exceed the
current length of the file, and instead use the default values.

Ref: LOG-17217
Signed-off-by: Jacob Hull <jacob@planethull.com>