fix: static builds failing due to rocksdb
Reverts rocksdb wrapper to 0.19.

This version of the agent is built on rocksdb 7.4.4 which is not
compatible with version 0.22 of the rocksdb wrapper. Since this version
of the agent is only receiving security updates and the previous
versions of rocksdb and the wrapper don't have any security issues,
let's just keep using the same version.

Signed-off-by: Jacob Hull <jacob@planethull.com>