fix: seat offset spanvec correctly after truncation
When truncating large files, the offsets SpanVec was incorrectly seeking
to the beginning of the file instead of the new offset (the end of the
file). Adds a new method to truncate `SpanVec` and utilize it during
file truncation.

Additionally `utils/stress_test` since it's not used, the previous
maintainer is not working on it anymore, and it's breaking our CI/CD
tooling.

Ref: LOG-21703
Signed-off-by: Jacob Hull <jacob@planethull.com>