fix: switch from winservice to windows_service
Our windows service utility crate is a 9 year old unmaintained project.
A previous contributor was using a fork of this crate that is also
unmaintained. Due to this and the fact we've been having CI/CD issues
because of git dependencies, we are swapping to the more comprehensive
and actively maintained `windows_service` crate.

Signed-off-by: Jacob Hull <jacob@planethull.com>